requirements_dev.txt
pip==19.2.3
bump2version==0.5.11
wheel==0.33.6
watchdog==0.9.0
flake8==3.7.8
tox==3.14.0
coverage==4.5.4
Sphinx==1.8.5
twine==1.14.0

pytest==4.6.5
pytest-runner==5.1

rich>=9.0.0
discord.py>=1.6
dice>=3.1
tinydb>=4.4
pyyaml>=5
xdg>=5
